Add Copy context menu for chatgpt widget
